fix! Address Chesnay's comment
