+++ flink-tests/src/test/java/org/apache/flink/test/state/operator/restore/unkeyed/ChainUnionTest.java
@@ -24,9 +24,6 @@ import org.apache.flink.test.state.operator.restore.ExecutionMode;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 
-import java.util.Arrays;
-import java.util.Collection;
-
 import static org.apache.flink.test.state.operator.restore.unkeyed.NonKeyedJob.createFirstStatefulMap;
 import static org.apache.flink.test.state.operator.restore.unkeyed.NonKeyedJob.createSecondStatefulMap;
 import static org.apache.flink.test.state.operator.restore.unkeyed.NonKeyedJob.createSource;
@@ -39,17 +36,8 @@ import static org.apache.flink.test.state.operator.restore.unkeyed.NonKeyedJob.c
 @RunWith(Parameterized.class)
 public class ChainUnionTest extends AbstractNonKeyedOperatorRestoreTestBase {
 
-	private final String savepointPath;
-
-	@Parameterized.Parameters(name = "Migrate Savepoint: {0}")
-	public static Collection<String> parameters () {
-		return Arrays.asList(
-			"nonKeyed-flink1.2",
-			"nonKeyed-flink1.3");
-	}
-
 	public ChainUnionTest(String savepointPath) {
-		this.savepointPath = savepointPath;
+		super(savepointPath);
 	}
 
 	@Override
@@ -68,10 +56,5 @@ public class ChainUnionTest extends AbstractNonKeyedOperatorRestoreTestBase {
 		SingleOutputStreamOperator<Integer> stateless = createStatelessMap(second);
 
 		SingleOutputStreamOperator<Integer> third = createThirdStatefulMap(ExecutionMode.RESTORE, stateless);
-	}
-
-	@Override
-	protected String getMigrationSavepointName() {
-		return savepointPath;
 	}
 }
